Add class constant NEW_LINE
Introduce a public constant to CLIPrintTool, so that users don't have to have prior knowledge to its exact string value.

I could've put the value of the constant as '\n', but the list() method technically doesn't print a new line, rather an empty string. So I kept its string value the same as the constant name 😄